[RF] Factor out likelihood and chi2 fitting code to `FitHelpers`
This is done so that we can easily reuse code between the likelihood and
chi2 fitting.

Also, the likelihood creation helpers from the former `BatchModeHelpers`
file was moved to `FitHelpers`, which is the only place where is was
used. It's not good to have a file called `BatchModeHelpers` if we don't
call it `BatchMode` anymore.
