+++ roofit/roofitcore/src/FitHelpers.cxx
@@ -1,3 +1,5 @@
+/// \cond ROOFIT_INTERNAL
+
 /*
  * Project: RooFit
  * Authors:
@@ -15,17 +17,41 @@
 #include <RooAbsData.h>
 #include <RooAbsPdf.h>
 #include <RooAbsReal.h>
+#include <RooAddition.h>
+#include <RooBatchCompute.h>
+#include <RooBinSamplingPdf.h>
+#include <RooCategory.h>
 #include <RooCmdConfig.h>
+#include <RooConstraintSum.h>
+#include <RooDataHist.h>
+#include <RooDataSet.h>
 #include <RooDerivative.h>
+#include <RooFit/Evaluator.h>
+#include <RooFit/TestStatistics/RooRealL.h>
+#include <RooFit/TestStatistics/buildLikelihood.h>
 #include <RooFitResult.h>
+#include <RooFuncWrapper.h>
 #include <RooLinkedList.h>
 #include <RooMinimizer.h>
 #include <RooRealVar.h>
+#include <RooSimultaneous.h>
 
 #include <Math/CholeskyDecomp.h>
 
-namespace RooFit {
-namespace FitHelpers {
+#include "ConstraintHelpers.h"
+#include "RooEvaluatorWrapper.h"
+#include "RooFitImplHelpers.h"
+#include "RooNLLVarNew.h"
+
+#ifdef ROOFIT_LEGACY_EVAL_BACKEND
+#include "RooChi2Var.h"
+#include "RooNLLVar.h"
+#include "RooXYChi2Var.h"
+#endif
+
+namespace {
+
+constexpr int extendedFitDefault = 2;
 
 ////////////////////////////////////////////////////////////////////////////////
 /// Use the asymptotically correct approach to estimate errors in the presence of weights.
@@ -155,8 +181,6 @@ int calcSumW2CorrectedCovariance(RooAbsReal const &pdf, RooMinimizer &minimizer,
    return std::min(rw->covQual(), rw2->covQual());
 }
 
-namespace {
-
 /// Configuration struct for RooAbsPdf::minimizeNLL with all the default values
 /// that also should be taked as the default values for RooAbsPdf::fitTo.
 struct MinimizerConfig {
@@ -186,7 +210,172 @@ struct MinimizerConfig {
    std::string minAlg = "minuit";
 };
 
+bool interpretExtendedCmdArg(RooAbsPdf const &pdf, int extendedCmdArg)
+{
+   // Process automatic extended option
+   if (extendedCmdArg == extendedFitDefault) {
+      bool ext = pdf.extendMode() == RooAbsPdf::CanBeExtended || pdf.extendMode() == RooAbsPdf::MustBeExtended;
+      if (ext) {
+         oocoutI(&pdf, Minimization)
+            << "p.d.f. provides expected number of events, including extended term in likelihood." << std::endl;
+      }
+      return ext;
+   }
+   // If Extended(false) was explicitly set, but the pdf MUST be extended, then
+   // it's time to print an error. This happens when you're fitting a RooAddPdf
+   // with coefficient that represent yields, and without the additional
+   // constraint these coefficients are degenerate because the RooAddPdf
+   // normalizes itself. Nothing correct can come out of this.
+   if (extendedCmdArg == 0) {
+      if (pdf.extendMode() == RooAbsPdf::MustBeExtended) {
+         std::string errMsg = "You used the Extended(false) option on a pdf where the fit MUST be extended! "
+                              "The parameters are not well defined and you're getting nonsensical results.";
+         oocoutE(&pdf, InputArguments) << errMsg << std::endl;
+      }
+   }
+   return extendedCmdArg;
+}
+
+/// To set the fitrange attribute of the PDF and custom ranges for the
+/// observables so that RooPlot can automatically plot the fitting range.
+void resetFitrangeAttributes(RooAbsArg &pdf, RooAbsData const &data, std::string const &baseName, const char *rangeName,
+                             bool splitRange)
+{
+   // Clear possible range attributes from previous fits.
+   pdf.removeStringAttribute("fitrange");
+
+   // No fitrange was specified, so we do nothing. Or "SplitRange" is used, and
+   // then there are no uniquely defined ranges for the observables (as they
+   // are different in each category).
+   if (!rangeName || splitRange)
+      return;
+
+   RooArgSet observables;
+   pdf.getObservables(data.get(), observables);
+
+   std::string fitrangeValue;
+   auto subranges = ROOT::Split(rangeName, ",");
+   for (auto const &subrange : subranges) {
+      if (subrange.empty())
+         continue;
+      std::string fitrangeValueSubrange = std::string("fit_") + baseName;
+      if (subranges.size() > 1) {
+         fitrangeValueSubrange += "_" + subrange;
+      }
+      fitrangeValue += fitrangeValueSubrange + ",";
+      for (RooAbsArg *arg : observables) {
+
+         if (arg->isCategory())
+            continue;
+         auto &observable = static_cast<RooRealVar &>(*arg);
+
+         observable.setRange(fitrangeValueSubrange.c_str(), observable.getMin(subrange.c_str()),
+                             observable.getMax(subrange.c_str()));
+      }
+   }
+   pdf.setStringAttribute("fitrange", fitrangeValue.substr(0, fitrangeValue.size() - 1).c_str());
+}
+
+std::unique_ptr<RooAbsArg> createSimultaneousNLL(RooSimultaneous const &simPdf, bool isExtended,
+                                                 std::string const &rangeName, RooFit::OffsetMode offset)
+{
+   RooAbsCategoryLValue const &simCat = simPdf.indexCat();
+
+   // Prepare the NLL terms for each component
+   RooArgList nllTerms;
+   for (auto const &catState : simCat) {
+      std::string const &catName = catState.first;
+      RooAbsCategory::value_type catIndex = catState.second;
+
+      // If the channel is not in the selected range of the category variable, we
+      // won't create an for NLL this channel.
+      if (!rangeName.empty()) {
+         // Only the RooCategory supports ranges, not the other
+         // RooAbsCategoryLValue-derived classes.
+         auto simCatAsRooCategory = dynamic_cast<RooCategory const *>(&simCat);
+         if (simCatAsRooCategory && !simCatAsRooCategory->isStateInRange(rangeName.c_str(), catIndex)) {
+            continue;
+         }
+      }
+
+      if (RooAbsPdf *pdf = simPdf.getPdf(catName.c_str())) {
+         auto name = std::string("nll_") + pdf->GetName();
+         std::unique_ptr<RooArgSet> observables(
+            static_cast<RooArgSet *>(std::unique_ptr<RooArgSet>(pdf->getVariables())->selectByAttrib("__obs__", true)));
+         auto nll = std::make_unique<RooNLLVarNew>(name.c_str(), name.c_str(), *pdf, *observables, isExtended, offset);
+         // Rename the special variables
+         nll->setPrefix(std::string("_") + catName + "_");
+         nllTerms.addOwned(std::move(nll));
+      }
+   }
+
+   for (auto *nll : static_range_cast<RooNLLVarNew *>(nllTerms)) {
+      nll->setSimCount(nllTerms.size());
+   }
+
+   // Time to sum the NLLs
+   auto nll = std::make_unique<RooAddition>("mynll", "mynll", nllTerms);
+   nll->addOwnedComponents(std::move(nllTerms));
+   return nll;
+}
+
+std::unique_ptr<RooAbsReal> createNLLNew(RooAbsPdf &pdf, RooAbsData &data, std::unique_ptr<RooAbsReal> &&constraints,
+                                         std::string const &rangeName, RooArgSet const &projDeps, bool isExtended,
+                                         double integrateOverBinsPrecision, RooFit::OffsetMode offset)
+{
+   if (constraints) {
+      // Redirect the global observables to the ones from the dataset if applicable.
+      constraints->setData(data, false);
+
+      // The computation graph for the constraints is very small, no need to do
+      // the tracking of clean and dirty nodes here.
+      constraints->setOperMode(RooAbsArg::ADirty);
+   }
+
+   RooArgSet observables;
+   pdf.getObservables(data.get(), observables);
+   observables.remove(projDeps, true, true);
+
+   oocxcoutI(&pdf, Fitting) << "RooAbsPdf::fitTo(" << pdf.GetName()
+                            << ") fixing normalization set for coefficient determination to observables in data"
+                            << "\n";
+   pdf.fixAddCoefNormalization(observables, false);
+
+   // Deal with the IntegrateBins argument
+   RooArgList binSamplingPdfs;
+   std::unique_ptr<RooAbsPdf> wrappedPdf = RooBinSamplingPdf::create(pdf, data, integrateOverBinsPrecision);
+   RooAbsPdf &finalPdf = wrappedPdf ? *wrappedPdf : pdf;
+   if (wrappedPdf) {
+      binSamplingPdfs.addOwned(std::move(wrappedPdf));
+   }
+   // Done dealing with the IntegrateBins option
+
+   RooArgList nllTerms;
+
+   auto simPdf = dynamic_cast<RooSimultaneous *>(&finalPdf);
+   if (simPdf) {
+      simPdf->wrapPdfsInBinSamplingPdfs(data, integrateOverBinsPrecision);
+      nllTerms.addOwned(createSimultaneousNLL(*simPdf, isExtended, rangeName, offset));
+   } else {
+      nllTerms.addOwned(
+         std::make_unique<RooNLLVarNew>("RooNLLVarNew", "RooNLLVarNew", finalPdf, observables, isExtended, offset));
+   }
+   if (constraints) {
+      nllTerms.addOwned(std::move(constraints));
+   }
+
+   std::string nllName = std::string("nll_") + pdf.GetName() + "_" + data.GetName();
+   auto nll = std::make_unique<RooAddition>(nllName.c_str(), nllName.c_str(), nllTerms);
+   nll->addOwnedComponents(std::move(binSamplingPdfs));
+   nll->addOwnedComponents(std::move(nllTerms));
+
+   return nll;
+}
+
 } // namespace
+
+namespace RooFit {
+namespace FitHelpers {
 
 void defineMinimizationOptions(RooCmdConfig &pc)
 {
@@ -332,9 +521,9 @@ std::unique_ptr<RooFitResult> minimize(RooAbsReal &pdf, RooAbsReal &nll, RooAbsD
 
    if (m.getNPar() > 0) {
       if (cfg.doAsymptotic == 1)
-         corrCovQual = RooFit::FitHelpers::calcAsymptoticCorrectedCovariance(pdf, m, data); // Asymptotically correct
+         corrCovQual = calcAsymptoticCorrectedCovariance(pdf, m, data); // Asymptotically correct
       if (cfg.doSumW2 == 1)
-         corrCovQual = RooFit::FitHelpers::calcSumW2CorrectedCovariance(pdf, m, nll);
+         corrCovQual = calcSumW2CorrectedCovariance(pdf, m, nll);
    }
 
    if (cfg.minos)
@@ -355,5 +544,519 @@ std::unique_ptr<RooFitResult> minimize(RooAbsReal &pdf, RooAbsReal &nll, RooAbsD
    return ret;
 }
 
+std::unique_ptr<RooAbsReal> createNLL(RooAbsPdf &pdf, RooAbsData &data, const RooLinkedList &cmdList)
+{
+   auto baseName = std::string("nll_") + pdf.GetName() + "_" + data.GetName();
+
+   // Select the pdf-specific commands
+   RooCmdConfig pc("RooAbsPdf::createNLL(" + std::string(pdf.GetName()) + ")");
+
+   pc.defineString("rangeName", "RangeWithName", 0, "", true);
+   pc.defineString("addCoefRange", "SumCoefRange", 0, "");
+   pc.defineString("globstag", "GlobalObservablesTag", 0, "");
+   pc.defineString("globssource", "GlobalObservablesSource", 0, "data");
+   pc.defineDouble("rangeLo", "Range", 0, -999.);
+   pc.defineDouble("rangeHi", "Range", 1, -999.);
+   pc.defineInt("splitRange", "SplitRange", 0, 0);
+   pc.defineInt("ext", "Extended", 0, extendedFitDefault);
+   pc.defineInt("numcpu", "NumCPU", 0, 1);
+   pc.defineInt("interleave", "NumCPU", 1, 0);
+   pc.defineInt("verbose", "Verbose", 0, 0);
+   pc.defineInt("optConst", "Optimize", 0, 0);
+   pc.defineInt("cloneData", "CloneData", 0, 2);
+   pc.defineSet("projDepSet", "ProjectedObservables", 0, nullptr);
+   pc.defineSet("cPars", "Constrain", 0, nullptr);
+   pc.defineSet("glObs", "GlobalObservables", 0, nullptr);
+   pc.defineInt("doOffset", "OffsetLikelihood", 0, 0);
+   pc.defineSet("extCons", "ExternalConstraints", 0, nullptr);
+   pc.defineInt("EvalBackend", "EvalBackend", 0, static_cast<int>(RooFit::EvalBackend::defaultValue()));
+   pc.defineDouble("IntegrateBins", "IntegrateBins", 0, -1.);
+   pc.defineMutex("Range", "RangeWithName");
+   pc.defineMutex("GlobalObservables", "GlobalObservablesTag");
+   pc.defineInt("ModularL", "ModularL", 0, 0);
+
+   // New style likelihoods define parallelization through Parallelize(...) on fitTo or attributes on
+   // RooMinimizer::Config.
+   pc.defineMutex("ModularL", "NumCPU");
+
+   // New style likelihoods define offsetting on minimizer, not on likelihood
+   pc.defineMutex("ModularL", "OffsetLikelihood");
+
+   // Process and check varargs
+   pc.process(cmdList);
+   if (!pc.ok(true)) {
+      return nullptr;
+   }
+
+   if (pc.getInt("ModularL")) {
+      int lut[3] = {2, 1, 0};
+      RooFit::TestStatistics::RooAbsL::Extended ext{
+         static_cast<RooFit::TestStatistics::RooAbsL::Extended>(lut[pc.getInt("ext")])};
+
+      RooArgSet cParsSet;
+      RooArgSet extConsSet;
+      RooArgSet glObsSet;
+
+      if (auto tmp = pc.getSet("cPars"))
+         cParsSet.add(*tmp);
+
+      if (auto tmp = pc.getSet("extCons"))
+         extConsSet.add(*tmp);
+
+      if (auto tmp = pc.getSet("glObs"))
+         glObsSet.add(*tmp);
+
+      const std::string rangeName = pc.getString("globstag", "", false);
+
+      RooFit::TestStatistics::NLLFactory builder{pdf, data};
+      builder.Extended(ext)
+         .ConstrainedParameters(cParsSet)
+         .ExternalConstraints(extConsSet)
+         .GlobalObservables(glObsSet)
+         .GlobalObservablesTag(rangeName.c_str());
+
+      return std::make_unique<RooFit::TestStatistics::RooRealL>("likelihood", "", builder.build());
+   }
+
+   // Decode command line arguments
+   const char *rangeName = pc.getString("rangeName", nullptr, true);
+   const char *addCoefRangeName = pc.getString("addCoefRange", nullptr, true);
+   const bool ext = interpretExtendedCmdArg(pdf, pc.getInt("ext"));
+
+   int splitRange = pc.getInt("splitRange");
+   int optConst = pc.getInt("optConst");
+   int cloneData = pc.getInt("cloneData");
+   auto offset = static_cast<RooFit::OffsetMode>(pc.getInt("doOffset"));
+
+   // If no explicit cloneData command is specified, cloneData is set to true if optimization is activated
+   if (cloneData == 2) {
+      cloneData = optConst;
+   }
+
+   if (pc.hasProcessed("Range")) {
+      double rangeLo = pc.getDouble("rangeLo");
+      double rangeHi = pc.getDouble("rangeHi");
+
+      // Create range with name 'fit' with above limits on all observables
+      RooArgSet obs;
+      pdf.getObservables(data.get(), obs);
+      for (auto arg : obs) {
+         RooRealVar *rrv = dynamic_cast<RooRealVar *>(arg);
+         if (rrv)
+            rrv->setRange("fit", rangeLo, rangeHi);
+      }
+
+      // Set range name to be fitted to "fit"
+      rangeName = "fit";
+   }
+
+   // Set the fitrange attribute of th PDF, add observables ranges for plotting
+   resetFitrangeAttributes(pdf, data, baseName, rangeName, splitRange);
+
+   RooArgSet projDeps;
+   auto tmp = pc.getSet("projDepSet");
+   if (tmp) {
+      projDeps.add(*tmp);
+   }
+
+   const std::string globalObservablesSource = pc.getString("globssource", "data", false);
+   if (globalObservablesSource != "data" && globalObservablesSource != "model") {
+      std::string errMsg = "RooAbsPdf::fitTo: GlobalObservablesSource can only be \"data\" or \"model\"!";
+      oocoutE(&pdf, InputArguments) << errMsg << std::endl;
+      throw std::invalid_argument(errMsg);
+   }
+   const bool takeGlobalObservablesFromData = globalObservablesSource == "data";
+
+   // Lambda function to create the correct constraint term for a PDF. In old
+   // RooFit, we use this PDF itself as the argument, for the new BatchMode
+   // we're passing a clone.
+   auto createConstr = [&](bool removeConstraintsFromPdf = false) -> std::unique_ptr<RooAbsReal> {
+      return createConstraintTerm(baseName + "_constr",                    // name
+                                  pdf,                                     // pdf
+                                  data,                                    // data
+                                  pc.getSet("cPars"),                      // Constrain RooCmdArg
+                                  pc.getSet("extCons"),                    // ExternalConstraints RooCmdArg
+                                  pc.getSet("glObs"),                      // GlobalObservables RooCmdArg
+                                  pc.getString("globstag", nullptr, true), // GlobalObservablesTag RooCmdArg
+                                  takeGlobalObservablesFromData,           // From GlobalObservablesSource RooCmdArg
+                                  removeConstraintsFromPdf);
+   };
+
+   auto evalBackend = static_cast<RooFit::EvalBackend::Value>(pc.getInt("EvalBackend"));
+
+   // Construct BatchModeNLL if requested
+   if (evalBackend != RooFit::EvalBackend::Value::Legacy) {
+
+      // Set the normalization range. We need to do it now, because it will be
+      // considered in `compileForNormSet`.
+      std::string oldNormRange;
+      if (pdf.normRange()) {
+         oldNormRange = pdf.normRange();
+      }
+      pdf.setNormRange(rangeName);
+
+      RooArgSet normSet;
+      pdf.getObservables(data.get(), normSet);
+      normSet.remove(projDeps, true, true);
+
+      pdf.setAttribute("SplitRange", splitRange);
+      pdf.setStringAttribute("RangeName", rangeName);
+
+      RooFit::Detail::CompileContext ctx{normSet};
+      ctx.setLikelihoodMode(true);
+      std::unique_ptr<RooAbsArg> head = pdf.compileForNormSet(normSet, ctx);
+      std::unique_ptr<RooAbsPdf> pdfClone = std::unique_ptr<RooAbsPdf>{static_cast<RooAbsPdf *>(head.release())};
+
+      // reset attributes
+      pdf.setAttribute("SplitRange", false);
+      pdf.setStringAttribute("RangeName", nullptr);
+
+      // Reset the normalization range
+      pdf.setNormRange(oldNormRange.c_str());
+
+      if (addCoefRangeName) {
+         oocxcoutI(&pdf, Fitting) << "RooAbsPdf::fitTo(" << pdf.GetName()
+                                  << ") fixing interpretation of coefficients of any component to range "
+                                  << addCoefRangeName << "\n";
+         pdfClone->fixAddCoefRange(addCoefRangeName, false);
+      }
+
+      std::unique_ptr<RooAbsReal> compiledConstr;
+      if (std::unique_ptr<RooAbsReal> constr = createConstr()) {
+         compiledConstr = RooFit::Detail::compileForNormSet(*constr, *data.get());
+         compiledConstr->addOwnedComponents(std::move(constr));
+      }
+
+      auto nll = createNLLNew(*pdfClone, data, std::move(compiledConstr), rangeName ? rangeName : "", projDeps, ext,
+                              pc.getDouble("IntegrateBins"), offset);
+
+      std::unique_ptr<RooAbsReal> nllWrapper;
+
+      if (evalBackend == RooFit::EvalBackend::Value::Codegen ||
+          evalBackend == RooFit::EvalBackend::Value::CodegenNoGrad) {
+         bool createGradient = evalBackend == RooFit::EvalBackend::Value::Codegen;
+         auto simPdf = dynamic_cast<RooSimultaneous const *>(pdfClone.get());
+         nllWrapper = std::make_unique<RooFuncWrapper>("nll_func_wrapper", "nll_func_wrapper", *nll, normSet, &data,
+                                                       simPdf, createGradient);
+      } else {
+         auto evaluator = std::make_unique<RooFit::Evaluator>(*nll, evalBackend == RooFit::EvalBackend::Value::Cuda);
+         nllWrapper = std::make_unique<RooEvaluatorWrapper>(*nll, std::move(evaluator), rangeName ? rangeName : "",
+                                                            dynamic_cast<RooSimultaneous *>(pdfClone.get()),
+                                                            takeGlobalObservablesFromData);
+         nllWrapper->setData(data, false);
+      }
+
+      nllWrapper->addOwnedComponents(std::move(nll));
+      nllWrapper->addOwnedComponents(std::move(pdfClone));
+      return nllWrapper;
+   }
+
+   std::unique_ptr<RooAbsReal> nll;
+
+#ifdef ROOFIT_LEGACY_EVAL_BACKEND
+   bool verbose = pc.getInt("verbose");
+
+   int numcpu = pc.getInt("numcpu");
+   int numcpu_strategy = pc.getInt("interleave");
+   // strategy 3 works only for RooSimultaneous.
+   if (numcpu_strategy == 3 && !pdf.InheritsFrom("RooSimultaneous")) {
+      oocoutW(&pdf, Minimization) << "Cannot use a NumCpu Strategy = 3 when the pdf is not a RooSimultaneous, "
+                                     "falling back to default strategy = 0"
+                                  << std::endl;
+      numcpu_strategy = 0;
+   }
+   RooFit::MPSplit interl = (RooFit::MPSplit)numcpu_strategy;
+
+   auto binnedLInfo = RooHelpers::getBinnedL(pdf);
+   RooAbsPdf &actualPdf = binnedLInfo.binnedPdf ? *binnedLInfo.binnedPdf : pdf;
+
+   // Construct NLL
+   RooAbsReal::setEvalErrorLoggingMode(RooAbsReal::CollectErrors);
+   RooAbsTestStatistic::Configuration cfg;
+   cfg.addCoefRangeName = addCoefRangeName ? addCoefRangeName : "";
+   cfg.nCPU = numcpu;
+   cfg.interleave = interl;
+   cfg.verbose = verbose;
+   cfg.splitCutRange = static_cast<bool>(splitRange);
+   cfg.cloneInputData = static_cast<bool>(cloneData);
+   cfg.integrateOverBinsPrecision = pc.getDouble("IntegrateBins");
+   cfg.binnedL = binnedLInfo.isBinnedL;
+   cfg.takeGlobalObservablesFromData = takeGlobalObservablesFromData;
+   cfg.rangeName = rangeName ? rangeName : "";
+   auto nllVar = std::make_unique<RooNLLVar>(baseName.c_str(), "-log(likelihood)", actualPdf, data, projDeps, ext, cfg);
+   nllVar->enableBinOffsetting(offset == RooFit::OffsetMode::Bin);
+   nll = std::move(nllVar);
+   RooAbsReal::setEvalErrorLoggingMode(RooAbsReal::PrintErrors);
+
+   // Include constraints, if any, in likelihood
+   if (std::unique_ptr<RooAbsReal> constraintTerm = createConstr()) {
+
+      // Even though it is technically only required when the computation graph
+      // is changed because global observables are taken from data, it is safer
+      // to clone the constraint model in general to reset the normalization
+      // integral caches and avoid ASAN build failures (the PDF of the main
+      // measurement is cloned too anyway, so not much overhead). This can be
+      // reconsidered after the caching of normalization sets by pointer is changed
+      // to a more memory-safe solution.
+      constraintTerm = RooHelpers::cloneTreeWithSameParameters(*constraintTerm, data.get());
+
+      // Redirect the global observables to the ones from the dataset if applicable.
+      constraintTerm->setData(data, false);
+
+      // The computation graph for the constraints is very small, no need to do
+      // the tracking of clean and dirty nodes here.
+      constraintTerm->setOperMode(RooAbsArg::ADirty);
+
+      auto orignll = std::move(nll);
+      nll = std::make_unique<RooAddition>((baseName + "_with_constr").c_str(), "nllWithCons",
+                                          RooArgSet(*orignll, *constraintTerm));
+      nll->addOwnedComponents(std::move(orignll), std::move(constraintTerm));
+   }
+
+   if (optConst) {
+      nll->constOptimizeTestStatistic(RooAbsArg::Activate, optConst > 1);
+   }
+
+   if (offset == RooFit::OffsetMode::Initial) {
+      nll->enableOffsetting(true);
+   }
+#else
+   throw std::runtime_error("RooFit was not built with the legacy evaluation backend");
+#endif
+
+   return nll;
+}
+
+std::unique_ptr<RooAbsReal> createChi2(RooAbsReal &real, RooDataHist &data, const RooLinkedList &cmdList)
+{
+#ifdef ROOFIT_LEGACY_EVAL_BACKEND
+   // Construct Chi2
+   RooAbsReal::setEvalErrorLoggingMode(RooAbsReal::CollectErrors);
+   std::string baseName = "chi2_" + std::string(real.GetName()) + "_" + data.GetName();
+
+   // Clear possible range attributes from previous fits.
+   real.removeStringAttribute("fitrange");
+
+   RooCmdConfig pc("RooChi2Var::RooChi2Var");
+   pc.defineInt("etype", "DataError", 0, (Int_t)RooDataHist::Auto);
+   pc.defineInt("extended", "Extended", 0, extendedFitDefault);
+   pc.defineInt("num_cpu", "NumCPU", 0, 1);
+   pc.defineInt("verbose", "Verbose", 0, 1);
+   pc.defineInt("split_range", "SplitRange", 0, 0);
+   pc.defineDouble("integrate_bins", "IntegrateBins", 0, -1);
+   pc.defineString("rangeName", "RangeWithName", 0, "", true);
+   pc.defineString("addCoefRange", "SumCoefRange", 0, "");
+   pc.allowUndefined();
+
+   pc.process(cmdList);
+   if (!pc.ok(true)) {
+      return nullptr;
+   }
+
+   bool extended = false;
+   if (auto pdf = dynamic_cast<RooAbsPdf const *>(&real)) {
+      extended = interpretExtendedCmdArg(*pdf, pc.getInt("extended"));
+   }
+
+   RooDataHist::ErrorType etype = static_cast<RooDataHist::ErrorType>(pc.getInt("etype"));
+
+   const char *rangeName = pc.getString("rangeName", nullptr, true);
+   const char *addCoefRangeName = pc.getString("addCoefRange", nullptr, true);
+
+   RooAbsTestStatistic::Configuration cfg;
+   cfg.rangeName = rangeName ? rangeName : "";
+   cfg.nCPU = pc.getInt("num_cpu");
+   cfg.interleave = RooFit::Interleave;
+   cfg.verbose = static_cast<bool>(pc.getInt("verbose"));
+   cfg.cloneInputData = false;
+   cfg.integrateOverBinsPrecision = pc.getDouble("integrate_bins");
+   cfg.addCoefRangeName = addCoefRangeName ? addCoefRangeName : "";
+   cfg.splitCutRange = static_cast<bool>(pc.getInt("split_range"));
+   auto chi2 = std::make_unique<RooChi2Var>(baseName.c_str(), baseName.c_str(), real, data, extended, etype, cfg);
+
+   RooAbsReal::setEvalErrorLoggingMode(RooAbsReal::PrintErrors);
+
+   return chi2;
+#else
+   throw std::runtime_error("createChi2() is not supported without the legacy evaluation backend");
+   return nullptr;
+#endif
+}
+
+std::unique_ptr<RooAbsReal> createChi2(RooAbsReal &real, RooDataSet &xydata, const RooLinkedList &cmdList)
+{
+#ifdef ROOFIT_LEGACY_EVAL_BACKEND
+   // Select the pdf-specific commands
+   RooCmdConfig pc("RooAbsReal::createChi2(" + std::string(real.GetName()) + ")");
+
+   pc.defineInt("integrate", "Integrate", 0, 0);
+   pc.defineObject("yvar", "YVar", 0, nullptr);
+   pc.defineString("rangeName", "RangeWithName", 0, "", true);
+   pc.defineInt("numcpu", "NumCPU", 0, 1);
+   pc.defineInt("interleave", "NumCPU", 1, 0);
+   pc.defineInt("verbose", "Verbose", 0, 0);
+
+   // Process and check varargs
+   pc.process(cmdList);
+   if (!pc.ok(true)) {
+      return nullptr;
+   }
+
+   // Decode command line arguments
+   bool integrate = pc.getInt("integrate");
+   RooRealVar *yvar = static_cast<RooRealVar *>(pc.getObject("yvar"));
+   const char *rangeName = pc.getString("rangeName", nullptr, true);
+   Int_t numcpu = pc.getInt("numcpu");
+   Int_t numcpu_strategy = pc.getInt("interleave");
+   // strategy 3 works only for RooSimultaneous.
+   if (numcpu_strategy == 3 && !real.InheritsFrom("RooSimultaneous")) {
+      oocoutW(&real, Minimization) << "Cannot use a NumCpu Strategy = 3 when the pdf is not a RooSimultaneous, "
+                                      "falling back to default strategy = 0"
+                                   << std::endl;
+      numcpu_strategy = 0;
+   }
+   RooFit::MPSplit interl = (RooFit::MPSplit)numcpu_strategy;
+   bool verbose = pc.getInt("verbose");
+
+   RooAbsTestStatistic::Configuration cfg;
+   cfg.rangeName = rangeName ? rangeName : "";
+   cfg.nCPU = numcpu;
+   cfg.interleave = interl;
+   cfg.verbose = verbose;
+   cfg.verbose = false;
+
+   std::string name = "chi2_" + std::string(real.GetName()) + "_" + xydata.GetName();
+
+   return std::make_unique<RooXYChi2Var>(name.c_str(), name.c_str(), real, xydata, yvar, integrate, cfg);
+#else
+   throw std::runtime_error("createChi2() is not supported without the legacy evaluation backend");
+#endif
+}
+
+std::unique_ptr<RooFitResult> fitTo(RooAbsPdf &pdf, RooAbsData &data, const RooLinkedList &cmdList)
+{
+   // Select the pdf-specific commands
+   RooCmdConfig pc("RooAbsPdf::fitTo(" + std::string(pdf.GetName()) + ")");
+
+   RooLinkedList fitCmdList(cmdList);
+   std::string nllCmdListString =
+      "ProjectedObservables,Extended,Range,"
+      "RangeWithName,SumCoefRange,NumCPU,SplitRange,Constrained,Constrain,ExternalConstraints,"
+      "CloneData,GlobalObservables,GlobalObservablesSource,GlobalObservablesTag,"
+      "EvalBackend,IntegrateBins,ModularL";
+
+   if (!cmdList.FindObject("ModularL") || static_cast<RooCmdArg *>(cmdList.FindObject("ModularL"))->getInt(0) == 0)
+      nllCmdListString += ",OffsetLikelihood";
+
+   RooLinkedList nllCmdList = pc.filterCmdList(fitCmdList, nllCmdListString.c_str());
+
+   pc.defineDouble("prefit", "Prefit", 0, 0);
+   defineMinimizationOptions(pc);
+
+   // Process and check varargs
+   pc.process(fitCmdList);
+   if (!pc.ok(true)) {
+      return nullptr;
+   }
+
+   // TimingAnalysis works only for RooSimultaneous.
+   if (pc.getInt("timingAnalysis") && !pdf.InheritsFrom("RooSimultaneous")) {
+      oocoutW(&pdf, Minimization) << "The timingAnalysis feature was built for minimization with RooSimultaneous "
+                                     "and is not implemented for other PDF's. Please create a RooSimultaneous to "
+                                     "enable this feature."
+                                  << std::endl;
+   }
+
+   // Decode command line arguments
+   double prefit = pc.getDouble("prefit");
+
+   if (prefit != 0) {
+      size_t nEvents = static_cast<size_t>(prefit * data.numEntries());
+      if (prefit > 0.5 || nEvents < 100) {
+         oocoutW(&pdf, InputArguments) << "PrefitDataFraction should be in suitable range."
+                                       << "With the current PrefitDataFraction=" << prefit
+                                       << ", the number of events would be " << nEvents << " out of "
+                                       << data.numEntries() << ". Skipping prefit..." << std::endl;
+      } else {
+         size_t step = data.numEntries() / nEvents;
+
+         RooDataSet tiny("tiny", "tiny", *data.get(), data.isWeighted() ? RooFit::WeightVar() : RooCmdArg());
+
+         for (int i = 0; i < data.numEntries(); i += step) {
+            const RooArgSet *event = data.get(i);
+            tiny.add(*event, data.weight());
+         }
+         RooLinkedList tinyCmdList(cmdList);
+         pc.filterCmdList(tinyCmdList, "Prefit,Hesse,Minos,Verbose,Save,Timer");
+         RooCmdArg hesse_option = RooFit::Hesse(false);
+         RooCmdArg print_option = RooFit::PrintLevel(-1);
+
+         tinyCmdList.Add(&hesse_option);
+         tinyCmdList.Add(&print_option);
+
+         pdf.fitTo(tiny, tinyCmdList);
+      }
+   }
+
+   RooCmdArg modularL_option;
+   if (pc.getInt("parallelize") != 0 || pc.getInt("enableParallelGradient") || pc.getInt("enableParallelDescent")) {
+      // Set to new style likelihood if parallelization is requested
+      modularL_option = RooFit::ModularL(true);
+      nllCmdList.Add(&modularL_option);
+   }
+
+   std::unique_ptr<RooAbsReal> nll{pdf.createNLL(data, nllCmdList)};
+
+   return RooFit::FitHelpers::minimize(pdf, *nll, data, pc);
+}
+
+std::unique_ptr<RooFitResult> chi2FitTo(RooAbsReal &real, RooDataHist &data, const RooLinkedList &cmdList)
+{
+   // Select the pdf-specific commands
+   RooCmdConfig pc("RooAbsPdf::chi2FitTo(" + std::string(real.GetName()) + ")");
+
+   // Pull arguments to be passed to chi2 construction from list
+   RooLinkedList fitCmdList(cmdList);
+
+   auto createChi2DataHistCmdArgs = "Range,RangeWithName,NumCPU,Optimize,IntegrateBins,ProjectedObservables,"
+                                    "AddCoefRange,SplitRange,DataError,Extended";
+   RooLinkedList chi2CmdList = pc.filterCmdList(fitCmdList, createChi2DataHistCmdArgs);
+
+   defineMinimizationOptions(pc);
+
+   // Process and check varargs
+   pc.process(fitCmdList);
+   if (!pc.ok(true)) {
+      return nullptr;
+   }
+
+   std::unique_ptr<RooAbsReal> chi2{real.createChi2(data, chi2CmdList)};
+   return RooFit::FitHelpers::minimize(real, *chi2, data, pc);
+}
+
+std::unique_ptr<RooFitResult> chi2FitTo(RooAbsReal &real, RooDataSet &xydata, const RooLinkedList &cmdList)
+{
+   // Select the pdf-specific commands
+   RooCmdConfig pc("RooAbsPdf::chi2FitTo(" + std::string(real.GetName()) + ")");
+
+   // Pull arguments to be passed to chi2 construction from list
+   RooLinkedList fitCmdList(cmdList);
+   auto createChi2DataSetCmdArgs = "YVar,Integrate,RangeWithName,NumCPU,Verbose";
+   RooLinkedList chi2CmdList = pc.filterCmdList(fitCmdList, createChi2DataSetCmdArgs);
+
+   defineMinimizationOptions(pc);
+
+   // Process and check varargs
+   pc.process(fitCmdList);
+   if (!pc.ok(true)) {
+      return nullptr;
+   }
+
+   std::unique_ptr<RooAbsReal> xychi2{real.createChi2(xydata, chi2CmdList)};
+   return RooFit::FitHelpers::minimize(real, *xychi2, xydata, pc);
+}
+
 } // namespace FitHelpers
 } // namespace RooFit
+
+/// \endcond
